Reduce workspace redundancy in tree loading
When loading a tree, the query was fetching information about the root
workspace, and the same information about each child workspace. Then for
each child workspace, the same query was recursively called, with each
child workspace becoming the "root workspace" for the next query. So,
except for the very first root workspace, each workspace was having its
information fetched twice. This commit eliminates this redundancy.