Reduce pointer-fetching redundancy in tree loading
Previously, when loading a tree, every workspace in the tree used the
same GraphQL query. Accordingly every tree engaged in the same process
to fetch its connected pointers. All that is needed for the tree,
however, is an array of the available pointers for the entire tree. A
more efficient way to obtain this array is to have two different GraphQL
queries. One for the root workspace and one for all of the other
workspaces. The GraphQL query for the root workspace fetches the
available pointers for the entire tree. Because this query has a bird's
eye view of things, it can avoid querying the database multiple times
for each pointer. The GraphQL query for the children workspaces, on
other hand, avoids pointers and just fetches non-pointer information.