Expand on loading msg for root workspace in tree
Now that a lot of the tree loading happens up front with the query
associated with the root workspace, I want to make sure users know that
they may have to wait a few seconds until they see the tree appearing on
the page. We can change the exact wording.