Be more selective about spacers and intervention
Before, all text nodes in a pointer export were given a spacer
at the start and end. Now, the first text node gets a spacer at the
start, and the last text node gets a spacer at the end. This is what we
need and no more. Also, before, any position within a pointer export
was considered potentially problematic for cursor navigation. Now, only
the start of the first text node, and the end of the last text node are
considered potentially problematic. Again, this is what we need and no
more.