Use aphrodite to style exported pointer
This is instead of styled-components, which was causing problems when
imported pointers were nested inside exported pointers. In particular,
prior to this commit, if pointer A was nested inside pointer B, and you
imported A inside exported pointer C, then you could expand A within C,
but you couldn't expand B. This bug doesn't arise with Aphrodite
though. In this component we determine the css on a case-by-case basis
inside the render method, so it's no surprise that not all React style
libraries can handle all aspects of this case smoothly.