Factor out some of the exported pointer logic
We need to ensure that when you're exporting a nested pointer
that what you've selected to export does not extend to the edge. This
is because a cursor that extends to the edge will pick up what comes
after in a way that means that the pointer won't be properly nested,
and you'll get the "split" pointer bug that I addressed previously. I
refactored by taking this logic out of the file and putting it in its
own function/file.