Change square bracket export order
Instead of left-to-right, outside-in, square bracket export now occurs
left-to-right, inside-out. This means the cursor will end the process
to the right of the rightmost, outermost export. In typical cases,
this is what the user will expect. In atypical cases (e.g., when the
bracket that balances things is inserted in a nested position), this
might not be exactly what the user expects, but it is intuitive enough
that I doubt it will cause any confusion.