First pass at basic automated scheduling
Schedule history is stored in-memory on the server. It would of course
make sense to integrate this with Sequelize, but this would be a pretty
big project since there isn't even a User model yet. Another
simplification is that all trees are considered eligible. Lastly,
currently the algorithm just gives you a random workspace in the tree
that you worked on least recently. Well, not literally random, just
whatever is first in the query. Next I'll add that constraints on
exactly which workspace within this tree it should give you, basing
these constraints on #151.