Change how to decide whether to show "Loading..."
When viewing a single workspace, clicking on the "Parent" button causes
`workspace.loading` to equal true, but doesn't get rid of
`workspace.workspace`. So the previous approach would have the old
workspace on the screen until the new  workspace fully loaded. For
workspaces that take a while to load, this would result in 5+ seconds
of nothing happening, leading a typical user to think their clicking
the "Parent" button didn't do anything. But if you tie the presentation
of the "Loading..." message to whether `workspace.loading` equals true,
then the "Loading..." messages replaces the old workspace immediately
when the user presses the "Parent" button.