Big refactor of tracking time spent
Now workspaces have a field tracking the amount of time spent working
on the workspace. This field gets updated whenever a user gets assigned
a workspace and is working on that workspace, regardless of whether
the tree is in "time budget" mode or not. When the tree *is* in time
budget mode, then updating the time spent also updates the allocated
budget. So, the tree UI will report the time spent on each workspace
(and each subtree) regardless of the mode under which this work
occurred. And these changes are backwards-compatible so old trees will
still display accurate information.

With respect to refactoring. I made a more general Timer component that
basically just handles the "ticking", and can be given an optional UI
if there is an associated visual component.