Reset server DB between tests
Initially I tried to reset the server DB between tests using Sequelize's CLI. Essentially, going into the API docker service and running `yarn run db:reset`. That approach proved to be extremely difficult to get working, because Postgres won't drop a database if there's a current connection, and the tests themselves are creating connections. I was eventually able to get it working with well-placed timeouts and some conditional logic of the form "if an error arises because there's another connection, then wait a few seconds and try to drop the DB again", but it was very inelegant and quite slow with all the waiting.

I instead to decided to create a test-only (guarded by process.node.NODE_ENV === "test") API route to reset the DB. I adopted this approach from the Cypress documentation. This works much, much better.