Big (tho just beginning) refactor of tree cards
Previously there was one massive component containing (1) conditional
logic concerning how the pending status of async queries affected
what to render, (2) all of the logic more "internal" to the component
such as event handling. I split this component into two, with the
outer component isolating the inner component from any async logic.
This dramatically reduced the complexity of connecting the state of
the inner component to the outcome of a query. Prior to this
refactor, doing this involved a mess of logic split between the
structure (to handle the case of children cards where the query value
starts off as definite might never change) and the componentDidUpdate
hook (to handle the root-level case where the query value always
starts off undefined and eventually changes to the fetched value).