Modify scheduling algorithm
Now the algorithm first eliminates workspaces that fall within the
minimum distance from a worked-on workspace, then sorts according to
distance from edge, then sorts according to max distance from a
worked-on workspace.

Also, if the user opts to find a suboptimal workspace, this no longer
includes stale workspaces, but just reduces the minimum acceptable
distance from a work-on workspace from 1 to 2.