Restructure answer drafts
Previously when a user submitted a response, this triggered a race
between the answer draft saving (triggered by the navigation away from
the workspace page) and the answer draft getting saved as the final
answer (triggered by the pressing of the "Done" button). Sometimes,
incorrectly, the penultimate version of the answer draft got saved as
the answer. Given the current frontend architecture, it is extremely
difficult to guarantee these events occur in the correct sequence. So
what I did was say that when viewing a workspace, any child who is
resolved displays its answer draft, whereas any child not resolved
displays its final answer. And only when a workspace goes from resolved
to unresolved does the answer draft get copied to the answer.