Get Firefox async code working in beforeunload
Don't know why wrapping this in setTimeout makes this work. Without this, if you're working on a workspace, and then direct your browser tab to some other address (e.g., www.google.com), the leaveWorkspace mutation never makes it to the server. With the setTimeout, however, the leaveWorkspace mutation does make it to the server. Might do a little independent research on this at some point.