Get the root workspace scratchpad working
I also needed to modify the snapshot logic. Previously, the snapshot
logic assumed every block in the Redux store belonged to the workspace
or a child of the workspace. Now, however, we have a block that
belongs to the root workspace. So we need to make sure the snapshot
logic is aware of this change.